requirements/dev.txt
# AWS
boto3>=1.28.0
aws-cdk-lib>=2.0.0
constructs>=10.0.0

# Frontend
streamlit>=1.24.0

# ML & Data Processing
transformers>=4.30.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Vector Store & Database
opensearch-py>=2.3.0

# Development Tools
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
python-dotenv>=1.0.0

# API & Web
fastapi>=0.100.0
uvicorn>=0.22.0
requests>=2.31.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.0 